file renaming: more consistent usage of `_` prefix
