+++ sopa/constants.py
