+++ sopa/io/reader/cosmx.py
@@ -16,7 +16,7 @@ from spatialdata import SpatialData
 from spatialdata.models import Image2DModel, Labels2DModel, PointsModel, ShapesModel, TableModel
 from spatialdata.transformations import Scale
 
-from ..._constants import SopaAttrs
+from ...constants import SopaAttrs
 from .utils import _deduplicate_names, _default_image_kwargs
 
 log = logging.getLogger(__name__)
