+++ sopa/io/report/generate.py
@@ -8,7 +8,7 @@ import seaborn as sns
 from anndata import AnnData
 from spatialdata import SpatialData
 
-from ..._constants import LOW_AVERAGE_COUNT, SopaKeys
+from ...constants import LOW_AVERAGE_COUNT, SopaKeys
 from ...utils import get_boundaries, get_intensities, get_spatial_image
 from .engine import CodeBlock, Columns, Image, Message, Paragraph, Root, Section, SubSection
