+++ sopa/patches/__init__.py
@@ -1,5 +1,5 @@
-from ._patches import Patches2D
+from .patches import Patches2D
 from .loader import TileLoader
-from ._factory import make_image_patches, make_transcript_patches
+from .factory import make_image_patches, make_transcript_patches
 from .cluster import cluster_embeddings
 from .infer import compute_embeddings
