+++ sopa/segmentation/__init__.py
@@ -1,5 +1,5 @@
 from .resolve import solve_conflicts, combine
-from ._stainings import StainingSegmentation
+from .stainings import StainingSegmentation
 from ._tissue import tissue, shapes_bounding_box
 from .methods import cellpose, baysor, custom_staining_based, comseg, stardist, proseg
 from . import methods
