+++ sopa/shapes/__init__.py
@@ -1,4 +1,4 @@
 from ._rasterize import rasterize
-from ._validate import ensure_polygon, to_valid_polygons
-from ._expand import expand_radius, remove_overlap
+from .validate import ensure_polygon, to_valid_polygons
+from .expand import expand_radius, remove_overlap
 from ._vectorize import _vectorize_mask, pixel_outer_bounds, vectorize
