+++ sopa/shapes/validate.py
