+++ sopa/spatial/__init__.py
@@ -1,4 +1,4 @@
-from ._build import spatial_neighbors
+from .build import spatial_neighbors
 from .distance import cells_to_groups, mean_distance, prepare_network
 from .join import assign_transcript_to_cell, sjoin
 from .morpho import geometrize_niches, niches_geometry_stats, vectorize_niches
