+++ sopa/spatial/morpho.py
@@ -8,9 +8,9 @@ from scipy.spatial import Delaunay
 from shapely.geometry import MultiPolygon, Polygon
 from spatialdata import SpatialData
 
-from .._constants import SopaKeys
-from ._build import _check_has_delaunay
+from ..constants import SopaKeys
 from ._graph import Component
+from .build import _check_has_delaunay
 
 log = logging.getLogger(__name__)
